refactor: split test_app.py into many files
