+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+
+### Testing
+
+-   test_app.py has been refactored into many files.
+
 ## [1.7.3] - 2023-12-15
 
 ### Bug Fixes
